Extract UInt64 index to new file
